pyproject.toml
[tool.poetry]
name = "thecoderbot"
version = "0.1.0"
description = ""
authors = ["Luis Antonio Correa Leyva <correaleyval@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.5"
requests = "^2.25.1"
python-dotenv = "^0.17.1"
pydantic = "^1.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
